refactor(lobby-serksa): break components into their own files
