+++ src/modules/lobby-serksa/components/create-sheets/SchoolCreateSheet.tsx
@@ -1,0 +1,147 @@
+"use client";
+
+import { InputFormField } from "@components/form";
+import { Button } from "@components/ui/button";
+import { Form, FormField } from "@components/ui/form";
+import { Label } from "@components/ui/label";
+import {
+  Sheet,
+  SheetContent,
+  SheetDescription,
+  SheetFooter,
+  SheetHeader,
+  SheetTitle,
+  SheetTrigger,
+} from "@components/ui/sheet";
+import { toast } from "@components/ui/use-toast";
+import { zodResolver } from "@hookform/resolvers/zod";
+import { axiosInstance } from "@modules/lobby-serksa/fetch";
+import { CreateSchoolsSchema } from "@modules/lobby-serksa/models";
+import React from "react";
+import { useFieldArray, useForm } from "react-hook-form";
+
+async function onCreateHandler(data: { schools: { name: string }[] }) {
+  return axiosInstance.post(`/admin/schools`, data.schools).then((res) => {
+    return res.data as { code: number; message: string };
+  });
+}
+
+export function SchoolCreateSheet({
+  children,
+  refetch,
+}: {
+  children: React.ReactNode;
+  refetch: () => void;
+}) {
+  const [isOpen, setIsOpen] = React.useState(false);
+  const form = useForm<{ schools: { name: string }[] }>({
+    resolver: zodResolver(CreateSchoolsSchema),
+    defaultValues: { schools: [{ name: "" }] },
+  });
+
+  const fieldArray = useFieldArray({
+    control: form.control,
+    name: "schools",
+  });
+
+  return (
+    <Form {...form}>
+      <Sheet open={isOpen} onOpenChange={setIsOpen}>
+        <SheetTrigger asChild>{children}</SheetTrigger>
+        <SheetContent className="lg:max-w-4xl">
+          <form
+            className="grid h-full grid-rows-[auto,1fr,auto,auto]"
+            onSubmit={form.handleSubmit(
+              (d) => {
+                onCreateHandler(d)
+                  .then((response) => {
+                    form.reset();
+                    refetch();
+                    setIsOpen(false);
+                    toast({
+                      title: "Success",
+                      description: response.message,
+                    });
+                  })
+                  .catch((errRes) => {
+                    toast({
+                      title: "Failed to create schools",
+                      description: errRes.response.data.message,
+                      variant: "destructive",
+                    });
+                  });
+              },
+              () => {
+                toast({
+                  title: "Invalid form data",
+                  description: "Problem with schools data",
+                  variant: "destructive",
+                });
+              },
+            )}
+          >
+            <SheetHeader>
+              <SheetTitle>Create schools</SheetTitle>
+              <SheetDescription>
+                Make changes to your profile here. Click save when you&apos;re
+                done.
+              </SheetDescription>
+            </SheetHeader>
+            <div className="my-2 flex flex-col gap-2 overflow-y-auto p-2 px-4">
+              {fieldArray.fields.map((field, index) => {
+                return (
+                  <div
+                    key={field.id}
+                    className="rounded border-2 p-2 px-4 pb-4"
+                  >
+                    <p className="ml-auto w-fit rounded bg-gray-300 px-2 text-sm">
+                      {field.id}
+                    </p>
+                    <Label>Name</Label>
+                    <FormField
+                      control={form.control}
+                      name={`schools.${index}.name`}
+                      render={({ field }) => (
+                        <InputFormField
+                          field={field}
+                          errorField={
+                            form.formState.errors
+                              ? form.formState.errors.schools?.[index]?.name
+                              : undefined
+                          }
+                          placeholder="Name"
+                        />
+                      )}
+                    />
+                  </div>
+                );
+              })}
+            </div>
+            <div className="flex gap-2">
+              <Button
+                type="button"
+                onClick={() => fieldArray.append({ name: "" })}
+              >
+                Add School
+              </Button>
+              {fieldArray.fields.length > 1 && (
+                <Button
+                  type="button"
+                  variant="secondary"
+                  onClick={() =>
+                    fieldArray.remove(fieldArray.fields.length - 1)
+                  }
+                >
+                  Remove Last School
+                </Button>
+              )}
+            </div>
+            <SheetFooter>
+              <Button type="submit">Create</Button>
+            </SheetFooter>
+          </form>
+        </SheetContent>
+      </Sheet>
+    </Form>
+  );
+}
