+++ src/modules/lobby-serksa/components/create-sheets/index.ts
@@ -1,0 +1,4 @@
+export * from "./SchoolCreateSheet";
+export * from "./MajorCreateSheet";
+export * from "./SubjectCreateSheet";
+export * from "./ProfessorCreateSheet";
