+++ src/modules/lobby-serksa/components/data-tables/school-data-table.tsx
@@ -21,161 +21,10 @@ import { useQuery } from "@tanstack/react-query";
 import { schoolColumns } from "@modules/lobby-serksa/components/data-tables";
 import { axiosInstance } from "@modules/lobby-serksa/fetch";
 import React from "react";
-import {
-  AlertDialog,
-  AlertDialogAction,
-  AlertDialogCancel,
-  AlertDialogContent,
-  AlertDialogDescription,
-  AlertDialogFooter,
-  AlertDialogHeader,
-  AlertDialogTitle,
-  AlertDialogTrigger,
-} from "@components/ui/alert-dialog";
-import {
-  Sheet,
-  SheetContent,
-  SheetDescription,
-  SheetFooter,
-  SheetHeader,
-  SheetTitle,
-  SheetTrigger,
-} from "@components/ui/sheet";
-import { useFieldArray, useForm } from "react-hook-form";
-import { CreateSchoolsSchema } from "@modules/lobby-serksa/models";
-import { zodResolver } from "@hookform/resolvers/zod";
-import { Form, FormField } from "@components/ui/form";
-import { InputFormField } from "@components/form";
-import { toast } from "@components/ui/use-toast";
-import { Label } from "@components/ui/label";
-import { AdvanceSearch, SearchBar } from "@components";
-import { Input } from "@components/ui/input";
-
-async function onCreateHandler(data: { schools: { name: string }[] }) {
-  return axiosInstance.post(`/admin/schools`, data.schools).then((res) => {
-    return res.data as { code: number; message: string };
-  });
-}
-
-function CreateShoolSheet({
-  children,
-  refetch,
-}: {
-  children: React.ReactNode;
-  refetch: () => void;
-}) {
-  const [isOpen, setIsOpen] = React.useState(false);
-  const form = useForm<{ schools: { name: string }[] }>({
-    resolver: zodResolver(CreateSchoolsSchema),
-    defaultValues: { schools: [{ name: "" }] },
-  });
-
-  const fieldArray = useFieldArray({
-    control: form.control,
-    name: "schools",
-  });
-
-  return (
-    <Form {...form}>
-      <Sheet open={isOpen} onOpenChange={setIsOpen}>
-        <SheetTrigger asChild>{children}</SheetTrigger>
-        <SheetContent className="lg:max-w-4xl">
-          <form
-            className="grid h-full grid-rows-[auto,1fr,auto,auto]"
-            onSubmit={form.handleSubmit(
-              (d) => {
-                onCreateHandler(d)
-                  .then((response) => {
-                    form.reset();
-                    refetch();
-                    setIsOpen(false);
-                    toast({
-                      title: "Success",
-                      description: response.message,
-                    });
-                  })
-                  .catch((errRes) => {
-                    toast({
-                      title: "Failed to create schools",
-                      description: errRes.response.data.message,
-                      variant: "destructive",
-                    });
-                  });
-              },
-              () => {
-                toast({
-                  title: "Invalid form data",
-                  description: "Problem with schools data",
-                  variant: "destructive",
-                });
-              },
-            )}
-          >
-            <SheetHeader>
-              <SheetTitle>Create schools</SheetTitle>
-              <SheetDescription>
-                Make changes to your profile here. Click save when you&apos;re
-                done.
-              </SheetDescription>
-            </SheetHeader>
-            <div className="my-2 flex flex-col gap-2 overflow-y-auto p-2 px-4">
-              {fieldArray.fields.map((field, index) => {
-                return (
-                  <div
-                    key={field.id}
-                    className="rounded border-2 p-2 px-4 pb-4"
-                  >
-                    <p className="ml-auto w-fit rounded bg-gray-300 px-2 text-sm">
-                      {field.id}
-                    </p>
-                    <Label>Name</Label>
-                    <FormField
-                      control={form.control}
-                      name={`schools.${index}.name`}
-                      render={({ field }) => (
-                        <InputFormField
-                          field={field}
-                          errorField={
-                            form.formState.errors
-                              ? form.formState.errors.schools?.[index]?.name
-                              : undefined
-                          }
-                          placeholder="Name"
-                        />
-                      )}
-                    />
-                  </div>
-                );
-              })}
-            </div>
-            <div className="flex gap-2">
-              <Button
-                type="button"
-                onClick={() => fieldArray.append({ name: "" })}
-              >
-                Add School
-              </Button>
-              {fieldArray.fields.length > 1 && (
-                <Button
-                  type="button"
-                  variant="secondary"
-                  onClick={() =>
-                    fieldArray.remove(fieldArray.fields.length - 1)
-                  }
-                >
-                  Remove Last School
-                </Button>
-              )}
-            </div>
-            <SheetFooter>
-              <Button type="submit">Create</Button>
-            </SheetFooter>
-          </form>
-        </SheetContent>
-      </Sheet>
-    </Form>
-  );
-}
+import { SearchBar } from "@components";
+import { SchoolCreateSheet } from "../create-sheets";
+import { SchoolAdvanceSearch } from "../advance-searchs";
+import { DeleteSelectButton } from "../DeleteSelectButton";
 
 function onDeleteSelectedHandler(ids: number[]) {
   const deleteBody = ids.map((id) => ({ id }));
@@ -198,8 +47,6 @@ export function SchoolDataTable() {
   const [searchId, setSearchId] = useQueryState("id:eq", parseAsInteger);
   const [total, setTotal] = React.useState(0);
   const [rowSelection, setRowSelection] = React.useState({});
-
-  const idInputRef = React.useRef<HTMLInputElement>(null);
 
   const { data, refetch, isLoading } = useQuery({
     queryKey: [
@@ -247,82 +94,24 @@ export function SchoolDataTable() {
             placeholder="Search name ..."
             defaultValue={search}
           />
-          <AdvanceSearch
-            title="Advance school search"
-            description="If you want to do a more specific search, you can use this feature."
-            items={[
-              <div className="flex flex-col gap-4" key="id-search">
-                <Label>
-                  ID{" "}
-                  <span className="rounded bg-gray-500 px-2 py-1 text-secondary">
-                    number only
-                  </span>
-                </Label>
-                <Input
-                  ref={idInputRef}
-                  placeholder="ID"
-                  defaultValue={searchId || undefined}
-                />
-              </div>,
-            ]}
-            onConfirm={() => {
-              if (!idInputRef.current) return;
-              setSearchId(+idInputRef.current.value || null);
+          <SchoolAdvanceSearch
+            defaultValues={{ id: searchId }}
+            onConfirm={({ id }) => {
+              setSearchId(id);
             }}
           />
         </div>
-        <AlertDialog>
-          {table.getIsSomeRowsSelected() || table.getIsAllRowsSelected() ? (
-            <>
-              <AlertDialogTrigger asChild>
-                <Button
-                  variant={"outline"}
-                  className="text-red-400 hover:text-red-500"
-                >
-                  Delete Selected
-                </Button>
-              </AlertDialogTrigger>
-              <AlertDialogContent>
-                <AlertDialogHeader>
-                  <AlertDialogTitle>Are you absolutely sure?</AlertDialogTitle>
-                  <AlertDialogDescription>
-                    This action cannot be undone. This will permanently delete
-                    the school records.
-                  </AlertDialogDescription>
-                </AlertDialogHeader>
-                <AlertDialogFooter>
-                  <AlertDialogCancel>Cancel</AlertDialogCancel>
-                  <AlertDialogAction
-                    onClick={() => {
-                      const ids = table
-                        .getSelectedRowModel()
-                        .rows.flatMap((r) => r.original.id);
-                      if (ids.length === 0) return;
-                      onDeleteSelectedHandler(ids)
-                        .then((res) => {
-                          toast({ title: res.data.message });
-                          refetch();
-                          table.resetRowSelection(false);
-                        })
-                        .catch((errRes) => {
-                          toast({
-                            title: "Error",
-                            description: errRes.response.data.message,
-                            variant: "destructive",
-                          });
-                        });
-                    }}
-                  >
-                    Continue
-                  </AlertDialogAction>
-                </AlertDialogFooter>
-              </AlertDialogContent>
-            </>
-          ) : null}
-        </AlertDialog>
-        <CreateShoolSheet refetch={() => refetch()}>
+        <DeleteSelectButton
+          isHidden={
+            !(table.getIsSomeRowsSelected() || table.getIsAllRowsSelected())
+          }
+          onDeleteSelectedHandler={onDeleteSelectedHandler}
+          refetch={refetch}
+          table={table}
+        />
+        <SchoolCreateSheet refetch={() => refetch()}>
           <Button>Create</Button>
-        </CreateShoolSheet>
+        </SchoolCreateSheet>
       </div>
       <div className="overflow-hidden rounded-lg border">
         <Table>
