+++ src/modules/lobby-serksa/components/index.ts
@@ -1,1 +1,2 @@
 export * from "./SearchPopover";
+export * from "./DeleteSelectButton";
